WORKS: see commit description
NEXT STEPS -------------------

See if it also works with `hatchling` as build backend

CURRENT STATUS  --------------

Outside container, build wheel package by running:

  ```shell
  python -m build --wheel
  ```

this will generate a `build` directory - which you can ignore - and a
`dist` directory where you'll find the wheel file.

Now spin up the container, and run:

  ```shell
  pip install ./dist/direnv_backup...<snip>.whl
  ```

This will create scripts in `~/.local/bin`. The issue I had was that my
`PATH` did not contain `~/.local/bin`, so the command was not visible to
me.